pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prismtm"
version = "0.1.0"
description = "A hierarchical task management system for software development projects"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "GenGrue", email = "dev.grue.studio@gmail.com" }]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.0.0",
    "packaging>=21.0",
    "PyYAML>=6.0",
    "jsonschema>=4.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
prsm = "prismtm.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/prism-task-manager"
Repository = "https://github.com/yourusername/prism-task-manager"
Issues = "https://github.com/yourusername/prism-task-manager/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
prism_task_manager = ["schemas/**/*.json", "schemas/migrations/*.py"]
